add stripchart for display
modified stripchart.py from MIT to use sockets to receive data from readpvthread1.py (which is a modification of the original readpv.py).  communication is via sockets where readpvthread1.py gets data from photovoltaic system, decodes it and (currently) sends the pvWatts data to the stripchart.  Scaling the data to the plot was  tricky but works (may be klugy and not pythonic).  comments welcome.